requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Task queue
celery==5.4.0
redis==5.0.4

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Config / Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
httpx==0.27.0

# PDF generation
weasyprint==62.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0

# Dev tooling
ruff==0.4.4
